+++ test/mocks/MockTradeService.sol
@@ -6,14 +6,14 @@ import { ITradeService } from "../../src/services/interfaces/ITradeService.sol";
 contract MockTradeService is ITradeService {
   struct IncreasePositionInputs {
     address _primaryAccount;
-    uint256 _subAccountId;
+    uint8 _subAccountId;
     uint256 _marketIndex;
     int256 _sizeDelta;
   }
 
   struct DecreasePositionInputs {
     address _account;
-    uint256 _subAccountId;
+    uint8 _subAccountId;
     uint256 _marketIndex;
     uint256 _positionSizeE30ToDecrease;
     // @todo - support take profit token
@@ -38,7 +38,7 @@ contract MockTradeService is ITradeService {
 
   function increasePosition(
     address _primaryAccount,
-    uint256 _subAccountId,
+    uint8 _subAccountId,
     uint256 _marketIndex,
     int256 _sizeDelta
   ) external {
@@ -55,7 +55,7 @@ contract MockTradeService is ITradeService {
 
   function decreasePosition(
     address _account,
-    uint256 _subAccountId,
+    uint8 _subAccountId,
     uint256 _marketIndex,
     uint256 _positionSizeE30ToDecrease,
     address _tpToken
@@ -71,12 +71,7 @@ contract MockTradeService is ITradeService {
     );
   }
 
-  function forceClosePosition(
-    address _account,
-    uint256 _subAccountId,
-    uint256 _marketIndex,
-    address _tpToken
-  ) external {
+  function forceClosePosition(address _account, uint8 _subAccountId, uint256 _marketIndex, address _tpToken) external {
     decreasePositionCallCount++;
   }
 }
